requirements.txt
fastapi==0.103.0
pydantic==2.3.0
uuid==1.30.0
firebase-admin==6.2.0
pytest==7.4.0
uvicorn==0.22.0
PyYAML==6.0